[Android/Feat] 스위치 API 연동 S12P31D103-206

See merge request s12-final/S12P31D103!153

+++ Android/app/src/main/java/com/hogumiwarts/lumos/MainActivity.kt
@@ -114,7 +114,7 @@ class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        val deviceId = intent.getIntExtra("deviceId", -1) // 기본값 -1
+        val deviceId = intent.getLongExtra("deviceId", -1L) // 기본값 -1
         val deviceType = intent.getStringExtra("deviceType") ?: ""
