+++ Android/app/src/main/java/com/hogumiwarts/lumos/ui/screens/control/airpurifier/AirpurifierIntent.kt
@@ -2,9 +2,9 @@ package com.hogumiwarts.lumos.ui.screens.control.airpurifier
 
 
 sealed class AirpurifierIntent {
-    data class LoadAirpurifierStatus(val deviceId: Int): AirpurifierIntent()
+    data class LoadAirpurifierStatus(val deviceId: Long): AirpurifierIntent()
 
-    data class ChangeAirpurifierPower(val deviceId: Int, val activated: Boolean): AirpurifierIntent()
+    data class ChangeAirpurifierPower(val deviceId: Long, val activated: Boolean): AirpurifierIntent()
 
-    data class ChangeAirpurifierFenMode(val deviceId: Int, val fanMode: String): AirpurifierIntent()
+    data class ChangeAirpurifierFenMode(val deviceId: Long, val fanMode: String): AirpurifierIntent()
 }
